refactor: 좌석과 등급 map 로직 Rank 클래스로 이동
