+++ domain/src/main/java/com/example/domain/model/model/Rank.kt
@@ -3,5 +3,18 @@ package com.example.domain.model.model
 import com.example.domain.model.price.Price
 
 enum class Rank(val price: Price) {
-    A(Price(12000)), S(Price(15000)), B(Price(10000))
+    A(Price(12000)), S(Price(15000)), B(Price(10000));
+
+    companion object {
+        fun map(row: Int): Rank {
+            return when (row) {
+                1, 2 -> B
+                3, 4 -> S
+                5 -> A
+                else -> throw IllegalArgumentException(CANT_MATCHING_RANK_ERROR)
+            }
+        }
+
+        const val CANT_MATCHING_RANK_ERROR = "해당 열에 해당하는 등급이 없습니다."
+    }
 }
